+++ backend/src/users/users.service.ts
@@ -1,11 +1,12 @@
-import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
-import { UserStatus } from './user.enums';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import { UserRole, UserStatus } from './user.enums';
 import { CreateUserDto } from './dto/create-user.dto';
 import * as bcrypt from "bcrypt";
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { Repository } from 'typeorm';
+import type { JwtUser } from 'src/auth/get-user.models';
 
 @Injectable()
 export class UsersService {
@@ -24,6 +25,15 @@ export class UsersService {
         return user;
     }
 
+    async findByEmailForLogin(email: string) {
+        const user = await this.userRepository
+                             .createQueryBuilder("user")
+                             .addSelect("user.password")
+                             .where("user.email = :email", { email })
+                             .getOne();
+        return user;
+    }
+
     async createUser(createUserDto: CreateUserDto): Promise<User> {
         const found = await this.userRepository.findOne({where: {email: createUserDto.email}});
         if(found) throw new ConflictException("email already exists");
@@ -38,20 +48,20 @@ export class UsersService {
         return user;
     }
 
-    async updateUser(id: string, updateUserDto: UpdateUserDto): Promise<User>{
+    async updateUser(user1: JwtUser, id: string, updateUserDto: UpdateUserDto): Promise<User>{
         const {name, phoneNo} = updateUserDto;
         const user = await this.getUserById(id);
+        if(user.id !== user1.id && user.role !== UserRole.ADMIN) throw new ForbiddenException()
         if(name) user.name = name;
         if(phoneNo) user.phoneNo = phoneNo.toString();
         await this.userRepository.save(user);
         return user;
     }
 
-    async deleteUser(id: string): Promise<void>{
-       const result = await this.userRepository.delete(id)
-        if (result.affected === 0) {
-            throw new NotFoundException(`User with ID "${id}" not found`);
-        }
+    async deleteUser(user1: JwtUser, id: string): Promise<void>{
+        const user = await this.getUserById(id);
+        if(user.id !== user1.id && user.role !== UserRole.ADMIN) throw new ForbiddenException()
+        await this.userRepository.delete(id)
     }
 
 }
